Separate classes into packages
Organized the project structure by separating classes into dedicated
packages for better maintainability and clarity.

- Moved all custom exceptions to the 'exception' package.
- Grouped task-related classes (Task, Deadline, ToDo, Event) into the
'task' package.
- Placed local storage logic into the 'storage' package.
- Separated user interface components into the 'ui' package.
- Parsed input handling classes into the 'parser' package.
- Moved JavaFX GUI launchers and controllers into the 'gui' package.
- Moved command into 'command' package, preparing for refactoring of
commands into separate classes.

This restructuring improves code readability, enforces logical
separation of concerns, and makes it easier to navigate and extend
the project in the future.
