+++ src/main/java/steven/exception/MissingStartAndEndTimeException.java
@@ -1,4 +1,4 @@
-package steven;
+package steven.exception;
 
 /**
  * Thrown when a user attempts to add an Event without specifying the start or end time.
