+++ src/main/java/steven/gui/Main.java
@@ -1,4 +1,4 @@
-package steven;
+package steven.gui;
 
 import java.io.IOException;
 
@@ -7,6 +7,7 @@ import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
+import steven.Steven;
 
 /**
  * A GUI for Duke using FXML.
