+++ src/main/java/steven/gui/MainWindow.java
@@ -1,4 +1,4 @@
-package steven;
+package steven.gui;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
@@ -7,6 +7,7 @@ import javafx.scene.control.TextField;
 import javafx.scene.image.Image;
 import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.VBox;
+import steven.Steven;
 
 import java.util.Objects;
