+++ src/main/java/steven/task/Event.java
@@ -1,4 +1,6 @@
-package steven;
+package steven.task;
+
+import steven.exception.InvalidDateAndTimeFormatException;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
