+++ src/main/java/steven/task/Task.java
@@ -1,4 +1,4 @@
-package steven;
+package steven.task;
 
 /**
  * Represents a task that can be marked as done or not done.
