+++ src/main/java/steven/task/ToDo.java
@@ -1,4 +1,4 @@
-package steven;
+package steven.task;
 
 /**
  * Represents a task that does not have a date or time attached to it.
