+++ src/test/java/steven/StorageTest.java
@@ -3,6 +3,8 @@ package steven;
 import java.util.ArrayList;
 
 import org.junit.jupiter.api.Test;
+import steven.storage.Storage;
+import steven.task.Task;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
