+++ src/test/java/steven/TaskListTest.java
@@ -3,6 +3,9 @@ package steven;
 import java.util.ArrayList;
 
 import org.junit.jupiter.api.Test;
+import steven.exception.StevenException;
+import steven.task.Task;
+import steven.task.TaskList;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.fail;
